Cleanup for ongoing maintenance, initial xref support, test updates
* `verify_riak_stats` - Observer removed from Riak.
* `replication2_upgrade` - Run in chunks, instead of one at a time.
* `pb_cipher_suites` - Dead code inside deprecation macro conditionals removed.

Several deletions, all of which have fallen into disrepair and are no longer relevant in 3.x+.
Most deleted files have been removed in community version 3.2, so they're removed here to get them out of our way:
* Old pre-3.x cruft
* `perf` harness and related modules
* Giddyup stuff
* Smoke test

`rebar.config` tightens up several option groups, with the eventual goal of clean `xref` passes, so the less old gunk is in the way, the better.

The `observer` module has been renamed to `rt_observer` to remove potential conflicts with the OTP module of the same name - it's not clear whether this module has ongoing value.

wday-contrib 1014 1050 1059 1548

Co-authored-by: Adrian Fraiha <aef-@users.noreply.github.com>
Co-authored-by: Ted Burghart <tburghart@users.noreply.github.com>
Co-authored-by: Fred Dushin <fadushin@users.noreply.github.com>