Fix further dialyzer errors
For verify_dollarkeyreap, it is arguable that the type on the query is wrong.  It is possible to search for keys with sibling_count > 0 (i.e. find all keys), although this isn't the purpose for find_keys - hence the type requiring pos_integer() not non_neg_integer().  Using object_stats is a cleaner way of getting the required result, so this modification has been made to clear the dialyzer error.