Nhse d34 nhskv.i33 token (#19)
Tests for https://github.com/OpenRiak/riak_kv/pull/35

Adds a functional test `verify_conditional_put_strong` which is added to the `kv_all` group.  Also a test for profiling the performance of conditional PUTs - `verify_conditional_put_perf`.  This is not expected to be part of release testing, and so is not included in any specific groups.