Nhse o34 orkv.i42 aae tree exclude (#9)
* Add aae_tree_exclude test

* Resolve xref/dialyzer issues

* Add tree_exclude test for tictac aae

* Nhse o32 orkv.i58 extras (#37)

* Further updates in support of stats changes

verify_vnode_polling uses a removed function - now switched to a maintained one

* Use new stats function (previous removed)

* leveled_reload_recalc

The default has changed to the non-backwards compatible `true`.  So it needs to be set to `true` in the test to allow for current -> previous.

* Temp update rebar.config

To update type for riak erlang client bucket properties

* Add xref check to GHA

* Update erlang.yml

Attempt to trigger the action

* Attempt to prompt GHA

Following change of default branch

* Simple word change to prompt GHA

* Add simple test of cache reset

* Update log text following review

* Extend aside in test for individual PUT option

* Update rebar.config