Nhse o34 orkv.i49 reapinfailure (#13)
* Alter test for new behaviour on failure

Now reaps will be done on primaries, and retried on secondaries .. rather than waiting for all primaries to be available

* Test reaps and failure with replication

Confirm that reaps will be applied to all available primaries, and have a deferred application on primaries once they become available.

This should resolve issues with clusters being out-of-sync when reaps are applied during a failure.

* Update tests/nextgenrepl_deletewithfailure.erl

Co-authored-by: Thomas Arts <thomas.arts@quviq.com>

---------

Co-authored-by: Thomas Arts <thomas.arts@quviq.com>